requirements.txt
# Temel kütüphaneler
langchain>=0.1.0
llama-index>=0.9.0
streamlit>=1.30.0
pydantic>=2.5.0

# Veri kaynakları için
arxiv>=2.0.0
PyGithub>=2.1.0
huggingface-hub>=0.20.0
beautifulsoup4>=4.12.0
requests>=2.31.0

# Veritabanı
chromadb>=0.4.18
sqlalchemy>=2.0.0

# LLM entegrasyonu
openai>=1.6.0

# Yardımcı araçlar
python-dotenv>=1.0.0
tqdm>=4.66.0
numpy>=1.26.0
pandas>=2.1.0
